added the code to the repo

+++ pipeline/model_utils/llama3_model.py
@@ -113,9 +113,9 @@ class Llama3Model(ModelBase):
 
         model = AutoModelForCausalLM.from_pretrained(
             model_path,
-            torch_dtype=dtype,
+            dtype=dtype,
             trust_remote_code=True,
-            device_map= "auto"
+            device_map="auto"
         ).eval()
 
         model.requires_grad_(False)
